project_devops/apps/Project-A/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
anthropic
langchain
langchain-openai
langchain-community
pinecone-client
pgvector
pytest
httpx
python-dotenv
